+++ src/device_info.cpp
@@ -139,7 +139,7 @@ using namespace tinytc;
 
 extern "C" {
 tinytc_status_t tinytc_core_info_generic_create(tinytc_core_info_t *info, int32_t register_space,
-                                                int32_t max_work_group_size, uint32_t sgs_size,
+                                                int32_t max_work_group_size, size_t sgs_size,
                                                 int32_t const *sgs) {
     if (info == nullptr || sgs == nullptr) {
         return tinytc_status_invalid_arguments;
@@ -230,7 +230,7 @@ tinytc_status_t tinytc_core_info_intel_create_from_name(tinytc_core_info_t *info
 
 tinytc_status_t tinytc_core_info_intel_create(tinytc_core_info_t *info, uint32_t ip_version,
                                               int32_t num_eus_per_subslice,
-                                              int32_t num_threads_per_eu, uint32_t sgs_size,
+                                              int32_t num_threads_per_eu, size_t sgs_size,
                                               int32_t const *sgs) {
     if (info == nullptr || sgs == nullptr) {
         return tinytc_status_invalid_arguments;
@@ -243,8 +243,8 @@ tinytc_status_t tinytc_core_info_intel_create(tinytc_core_info_t *info, uint32_t
     });
 }
 
-tinytc_status_t tinytc_core_info_get_subgroup_sizes(const_tinytc_core_info_t info,
-                                                    uint32_t *sgs_size, int32_t const **sgs) {
+tinytc_status_t tinytc_core_info_get_subgroup_sizes(const_tinytc_core_info_t info, size_t *sgs_size,
+                                                    int32_t const **sgs) {
 
     if (info == nullptr || sgs_size == nullptr || sgs == nullptr) {
         return tinytc_status_invalid_arguments;
